requirements.txt
# Data Ingestion and Streaming
kafka-python==2.0.2

# DBT (Data Build Tool)
dbt-core==1.4.2
dbt-snowflake==1.4.2

# Apache Airflow
apache-airflow==2.5.0

# Machine Learning Libraries
scikit-learn==1.2.2
pandas==2.0.3
numpy==1.23.3
joblib==1.3.0

# Snowflake Connector for Python
snowflake
snowflake-connector-python==3.0.3
snowflake-sqlalchemy==1.4.7

# BI and Data Visualization (optional integrations)
# Uncomment if Looker is needed, otherwise use solution 2
looker-sdk>=22.0.0  # Using a newer Looker SDK version
# tableauserverclient==0.16

# Logging and Monitoring
prometheus_client==0.14.1

# Security
cryptography==40.0.0  # Version compatible with Snowflake and Airflow

# Additional Dependencies
psycopg2==2.9.6   # If you're using PostgreSQL or other databases for logging or staging
requests==2.31.0  # For APIs
sqlalchemy==1.4.46  # Compatible version for both Snowflake and Airflow

# Optional for Deployment and Containerization
docker==6.0.1
